Separated files out a little
